refactor: move articles, billing and dashboard components to apps/app
