+++ apps/app/components/articles/create-form.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import { buttonVariants } from "@/components/ui/button";
-import { toast } from "@/components/ui/sonner";
 import { useProject } from "@/hooks/use-project-context";
 import { type ArticleVariant } from "@/hooks/use-variant-operations";
-import { removeArticleCoverImage, updateArticle, updateArticleCoverImage } from "@/lib/actions/articles";
+import { createArticle, updateArticleCoverImage } from "@/lib/actions/articles";
 import { type LanguageCode, getLanguageName } from "@/lib/types/languages";
-import { X } from "lucide-react";
+import { buttonVariants } from "@simplist/ui/components/button";
+import { toast } from "@simplist/ui/components/sonner";
+import { Trash2 } from "lucide-react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
@@ -18,85 +18,31 @@ import { ArticleVisibilityCard } from "./visibility-card";
 
 type ArticleStatus = "draft" | "published" | "scheduled";
 
-type Article = {
-  id: string;
-  title: string;
-  excerpt: string | null;
-  content: string;
-  status: string;
-  coverImage: string | null;
+type CreateArticleFormProps = {
   projectId: string;
-  scheduledPublishAt?: Date | null;
-  variants?: Array<{
-    id: string;
-    lang: string;
-    title: string;
-    excerpt: string | null;
-    content: string;
-    coverImage: string | null;
-  }>;
-  project: {
-    defaultLanguage?: string;
-  };
 };
 
-type EditArticleFormProps = {
-  article: Article;
-};
-
-export const EditArticleForm = ({ article }: EditArticleFormProps) => {
+export const CreateArticleForm = ({ projectId }: CreateArticleFormProps) => {
   const router = useRouter();
   const { currentProject } = useProject();
   const [isSubmitting, setIsSubmitting] = useState(false);
-  const [isRemovingImage, setIsRemovingImage] = useState(false);
 
   // Default language from project or fallback to English
-  const defaultLanguage: LanguageCode = (article.project.defaultLanguage as LanguageCode) || (currentProject?.defaultLanguage as LanguageCode) || "en";
-
-  // Initialize variants with main article data and existing variants
-  const initializeVariants = (): ArticleVariant[] => {
-    const variants: ArticleVariant[] = [];
-    
-    // Add main article as default language variant
-    variants.push({
-      lang: defaultLanguage,
-      title: article.title,
-      excerpt: article.excerpt || "",
-      content: article.content,
-      coverImage: article.coverImage || undefined,
-    });
-
-    // Add existing variants (if any)
-    if (article.variants) {
-      article.variants.forEach(variant => {
-        if (variant.lang !== defaultLanguage) {
-          variants.push({
-            lang: variant.lang as LanguageCode,
-            title: variant.title,
-            excerpt: variant.excerpt || "",
-            content: variant.content,
-            coverImage: variant.coverImage || undefined,
-          });
-        }
-      });
-    }
-
-    return variants;
-  };
+  const defaultLanguage: LanguageCode = (currentProject?.defaultLanguage as LanguageCode) || "en";
 
   // Form state
-  const [title, setTitle] = useState(article.title);
-  const [excerpt, setExcerpt] = useState(article.excerpt || "");
-  const [content, setContent] = useState(article.content);
-  const [status, setStatus] = useState<ArticleStatus>(article.status as ArticleStatus);
-  const [scheduledPublishAt, setScheduledPublishAt] = useState<Date | null>(
-    article.scheduledPublishAt ? new Date(article.scheduledPublishAt) : null
-  );
-  const [imagePreview, setImagePreview] = useState<string | null>(article.coverImage);
+  const [title, setTitle] = useState("");
+  const [excerpt, setExcerpt] = useState("");
+  const [content, setContent] = useState("");
+  const [status, setStatus] = useState<ArticleStatus>("draft");
+  const [scheduledPublishAt, setScheduledPublishAt] = useState<Date | null>(null);
+  const [imagePreview, setImagePreview] = useState<string | null>(null);
   const [imageFile, setImageFile] = useState<File | null>(null);
 
     // Variants state
-    const [variants, setVariants] = useState<ArticleVariant[]>(initializeVariants());
+    const [variants, setVariants] = useState<ArticleVariant[]>([
+      { lang: defaultLanguage, title: "", excerpt: "", content: "" }
+    ]);
     const [activeVariant, setActiveVariant] = useState<LanguageCode>(defaultLanguage);
 
   // Sync form fields with active variant (default or selected)
@@ -139,30 +85,13 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
   };
 
   // Remove image
-  const handleRemoveImage = async () => {
-    setIsRemovingImage(true);
-    
-    // If there's a server image, delete it
-    if (article.coverImage && !imageFile) {
-      try {
-        await removeArticleCoverImage(article.id);
-        router.refresh();
-      } catch (error) {
-        console.error("Failed to remove cover image:", error);
-        toast.error("Failed to remove image");
-        setIsRemovingImage(false);
-        return;
-      }
-    }
-    
+  const handleRemoveImage = () => {
     setImagePreview(null);
     setImageFile(null);
     updateActiveVariant({ coverImage: undefined });
     // Reset file input
     const input = document.getElementById("image-upload") as HTMLInputElement;
     if (input) input.value = "";
-    
-    setIsRemovingImage(false);
   };
 
   // Handle form submission
@@ -184,11 +113,11 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
       }
     }
 
-    const toastId = toast.loading("Updating article...");
+    const toastId = toast.loading("Creating article...");
 
     try {
-      // Step 1: Update article
-      toast.loading("Updating article content and metadata...", { id: toastId });
+      // Step 1: Create article
+      toast.loading("Generating slug and calculating stats...", { id: toastId });
       
       // Get default language variant for main article
       const defaultVariant = variants.find(v => v.lang === defaultLanguage);
@@ -207,18 +136,20 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
           coverImage: v.coverImage,
         }));
 
-      await updateArticle(article.id, {
+      const article = await createArticle({
         title: defaultVariant.title,
         excerpt: defaultVariant.excerpt,
         content: defaultVariant.content,
         status,
-        scheduledPublishAt: status === "scheduled" ? scheduledPublishAt : null,
-        variants: articleVariants,
+        coverImage: defaultVariant.coverImage,
+        scheduledPublishAt: status === "scheduled" ? scheduledPublishAt || undefined : undefined,
+        projectId: currentProject?.id,
+        variants: articleVariants.length > 0 ? articleVariants : undefined,
       });
 
-      // Step 2: Upload new image if provided
-      if (imageFile) {
-        toast.loading("Uploading new cover image...", { id: toastId });
+      // Step 2: Upload image if provided
+      if (imageFile && article) {
+        toast.loading("Uploading cover image...", { id: toastId });
         const form = new FormData()
         form.append("file", imageFile)
         form.append("projectId", article.projectId)
@@ -243,13 +174,13 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
       }
 
       // Step 3: Success
-      toast.success("Article updated successfully!", { id: toastId });
+      toast.success("Article created successfully!", { id: toastId });
 
       router.push(`/${currentProject?.slug}/articles`);
       router.refresh();
     } catch (error) {
-      console.error("Error updating article:", error);
-      toast.error("Failed to update article. Please try again.", { id: toastId });
+      console.error("Error creating article:", error);
+      toast.error("Failed to create article. Please try again.", { id: toastId });
       setIsSubmitting(false);
     }
   };
@@ -287,7 +218,7 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
             status={status}
             onStatusChange={(v) => setStatus(v)}
             isSubmitting={isSubmitting}
-            submitLabel="Update"
+            submitLabel="Publish"
             scheduledPublishAt={scheduledPublishAt}
             onScheduleChange={setScheduledPublishAt}
             projectTimezone="UTC"
@@ -295,10 +226,10 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
             leftAction={(
               <Link
                 href={`/${currentProject?.slug}/articles`}
-                className={buttonVariants({ variant: "outline", size: "sm" })}
+                className={buttonVariants({ variant: "destructive", size: "sm" })}
               >
-                <X />
-                Cancel
+                <Trash2 />
+                Delete
               </Link>
             )}
           />
@@ -307,14 +238,14 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
             imagePreview={imagePreview}
             onImageChange={handlePickedImage}
             onRemoveImage={handleRemoveImage}
-            uploadLabel={activeVariant === defaultLanguage ? "Change Image" : `Change Image for ${getLanguageName(activeVariant)}`}
+            uploadLabel={activeVariant === defaultLanguage ? "Upload Image" : `Upload Image for ${getLanguageName(activeVariant)}`}
             emptyDescription={
               activeVariant === defaultLanguage 
-                ? "Update the article cover image."
+                ? "On the response API it will return the URL of the image."
                 : `Upload a specific image for ${getLanguageName(activeVariant)} variant. Each variant can have its own image.`
             }
-            isRemoving={isRemovingImage}
           />
+          
 
           <VariantCard
             defaultLanguage={defaultLanguage}
@@ -329,4 +260,3 @@ export const EditArticleForm = ({ article }: EditArticleFormProps) => {
     </form>
   );
 }
-
