+++ apps/app/components/articles/variant-card-base.tsx
