+++ apps/app/components/billing/current-plan-card.tsx
@@ -1,14 +1,14 @@
 "use client"
 
 import { UpgradeModal } from "@/components/billing/upgrade-modal"
-import { Badge } from "@/components/ui/badge"
-import { Button } from "@/components/ui/button"
-import { Card, CardContent } from "@/components/ui/card"
-import { Field, FieldContent, FieldDescription, FieldGroup, FieldLabel, FieldSeparator, FieldSet, FieldTitle } from "@/components/ui/field"
-import { Spinner } from "@/components/ui/spinner"
 import { createBillingPortalSession } from "@/lib/stripe/actions"
 import { getPlan } from "@/lib/subscription/plans"
 import { SubscriptionTier } from "@simplist/db/types"
+import { Badge } from "@simplist/ui/components/badge"
+import { Button } from "@simplist/ui/components/button"
+import { Card, CardContent } from "@simplist/ui/components/card"
+import { Field, FieldContent, FieldDescription, FieldGroup, FieldLabel, FieldSeparator, FieldSet, FieldTitle } from "@simplist/ui/components/field"
+import { Spinner } from "@simplist/ui/components/spinner"
 import { Check, CircleGauge } from "lucide-react"
 import { useRouter } from "next/navigation"
 import { useState } from "react"
