+++ apps/app/components/billing/upgrade-modal.tsx
@@ -1,18 +1,18 @@
 "use client"
 
-import { Badge } from "@/components/ui/badge"
-import { Button } from "@/components/ui/button"
-import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog"
-import { Field, FieldContent, FieldDescription, FieldGroup, FieldLabel, FieldSet, FieldTitle } from "@/components/ui/field"
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
 import { createCheckoutSession } from "@/lib/stripe/actions"
 import { getPlan } from "@/lib/subscription/plans"
 import { SubscriptionTier } from "@simplist/db/types"
+import { Badge } from "@simplist/ui/components/badge"
+import { Button } from "@simplist/ui/components/button"
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from "@simplist/ui/components/dialog"
+import { Field, FieldContent, FieldDescription, FieldGroup, FieldLabel, FieldSet, FieldTitle } from "@simplist/ui/components/field"
+import { RadioGroup, RadioGroupItem } from "@simplist/ui/components/radio-group"
+import { Spinner } from "@simplist/ui/components/spinner"
 import { Check, CircleFadingArrowUp } from "lucide-react"
 import { useRouter } from "next/navigation"
 import { useState } from "react"
 import { toast } from "sonner"
-import { Spinner } from "../ui/spinner"
 
 type UpgradeModalProps = {
   projectId: string
